Fix bugprone-implicit-widening-of-multiplication-result
warning: performing an implicit widening conversion to type 'size_t' (aka 'unsigned long') of a multiplication performed in type 'LyrIndex' (aka 'unsigned int')
warning: performing an implicit widening conversion to type 'unsigned long' of a multiplication performed in type 'int'

+++ src/SW_netCDF.c
@@ -1307,7 +1307,7 @@ static void alloc_netCDF_domain_vars(
             numVals = (bndVarNum % 2 == 0) ? numY : numX;
 
             *(bndsVars[bndVarNum]) = (double *) Mem_Malloc(
-                numVals * numBnds * sizeof(double),
+                (unsigned long) (numVals * numBnds) * sizeof(double),
                 "alloc_netCDF_domain_vars()",
                 LogInfo
             );
